Isolate equality I/O contract metrics
